Add a web interface
This will be useful for multiple reasons. We can change the generated
Bugzilla messages and show relevant links in the web UI, we can point
users of Fedora Infra folks to the messages page to know what we
received or not, we can use it to have always up-to-date statistics,
etc.

But I am mainly implementing it because I wanted to build something on
top of HTMX.